添加 测试数据方法 初始化数据方法

+++ ConsoleApplication1/test_k.cpp
@@ -13,6 +13,10 @@ TestK::TestK(int tmp){
 	k = tmp;
 }
 
+// 获得参数k的值
+int TestK::GetK(){
+	return k;
+}
 // 向数据库中插入数据
 int TestK::DataInsert(){
 
@@ -66,8 +70,41 @@ int TestK::DataInsert(){
 
 }
 
+bool TestK::TestData(){
+	string sql = "";
+	// 获取数据库中的第一条数据
+	sql = "select top 1 * from Tags";
+	adoConn.GetRecordSet(sql.c_str());
+	if (adoConn.m_pRecordset->GetState() == adStateClosed || adoConn.m_pRecordset->GetRecordCount() == 0)
+		return FALSE;
+	unint qua = adoConn.m_pRecordset->GetCollect("QuasiId").uintVal;
+	adoConn.CloseRecordset();
+
+	// 获取具有相同QuasiId的记录
+	sql = "select * from Tags where [QuasiId]=";
+	sql += itostr(qua);
+	adoConn.GetRecordSet(sql.c_str());
+	if (adoConn.m_pRecordset->GetState() == adStateClosed || adoConn.m_pRecordset->GetRecordCount() == 0)
+		return FALSE;
+	
+	// QuasiId相同的记录条数大于等于k值，则返回TRUE
+	if (adoConn.m_pRecordset->GetRecordCount() >= k){
+		adoConn.CloseRecordset();
+		return TRUE;
+	}
+	else{
+		adoConn.CloseRecordset();
+		return FALSE;
+	}
+}
+
 // 初始化数据库中的数据
 bool TestK::InitData(){
+	// 如果数据正确，则返回TRUE。
+	if (TestData())
+		return TRUE;
+
+	// 初始化数据
 	cout << endl << "Delete All Data..." << endl;
 	string sql = "";
 	sql = "TRUNCATE TABLE [Tags]";
@@ -97,8 +134,7 @@ std::ostream& operator<<(ostream & out, TestK::Response_Info value){		// 输出 响
 
 // 模拟标签行为的函数
 TestK::Response_Info TestK::TagFun(TestK::Request_Info req){
-	//std::cout << std::endl << "Request_Info: " << req << endl;
-	//std::cout << endl << "Start TagFun ..." << endl;	
+
 	// 读取 Request_Info内的信息
 	unint quaId, key, r1;
 	quaId = req.QuasiId;	
@@ -109,19 +145,14 @@ TestK::Response_Info TestK::TagFun(TestK::Request_Info req){
 	// 生成随机数，使用r1与key的并做为种子
 	minstd_rand0 minrand(r1&key);		
 	unint r2 = minrand();
-
-	//std::cout << "r1: " << r1 << "\tr2: " << r2 << endl;
-	//std::cout << "Key: " << key << endl;
 	
 	unint h = int_hash(r1&r2&key);	// 计算Hash值
-	//std::cout << "hash(r1,r2,key): " << h << endl;
 
 	Response_Info res;			// 生成响应信息
 	res.QuasiId = quaId;
 	res.r2 = r2;
 	res.h = h;
-	//std::cout << endl << "Response_Info: " << res << endl;
-	//std::cout << endl << "TagFun end." << endl;
+
 	return  res;
 }
 
@@ -135,16 +166,13 @@ TestK::Result_Info TestK::ReaderFun(Response_Info res, unint r1){
 	unint key, h1;		// 数据库中存储的标签密钥, 新的Hash值
 	hash<unint> int_hash;		// 计算Hash值
 
-	//cout << endl << "Start ReaderFun ... QuasiId = " << qua << endl;
-	// 访问数据库
-	//ADOConn adoConn;
 	string sql;
 	sql = "select [QuasiId],[Key] from Tags where QuasiId = ";
 	sql += itostr(qua);
 	_RecordsetPtr pRecordset = adoConn.GetRecordSet(sql.c_str());
 	
-	if (pRecordset->GetRecordCount() == 0)
-		return false;
+	if (pRecordset->GetState() == adStateClosed || pRecordset->GetRecordCount() == 0)
+		return FALSE;
 	// 遍历记录集
 	while (pRecordset->adoEOF == 0){
 		key = pRecordset->GetCollect("Key").uintVal;
@@ -156,8 +184,6 @@ TestK::Result_Info TestK::ReaderFun(Response_Info res, unint r1){
 		}
 		pRecordset->MoveNext();
 	}
-
-	//cout << endl << "ReaderFun end." << endl;
 	return FALSE;
 }
 
@@ -173,8 +199,6 @@ unsigned _int64 TestK::ProtocolFun(const int & tag){
 		return 0;
 	}
 
-	//cout << endl << "Start testing k-anonymous model...." << endl;
-	//cout << "\tTag[" << tag << "]" << endl;
 	// 生成查询信息
 	Request_Info ri;
 	ri.QuasiId = adoConn.m_pRecordset->GetCollect("QuasiId").uintVal;
@@ -188,31 +212,17 @@ unsigned _int64 TestK::ProtocolFun(const int & tag){
 	timer.Start();
 
 	// 生成随机数
-	//random_device rd;	
 	unint R1 = minrand();
-
-	//hash<string> str_hash;
-	//unint qua = str_hash("QuasiID of Tag");
-	//hash<unint> int_hash;
-	//unint key = int_hash(qua&r1);
-
 	ri.r1 = R1;
 
 	// 调用模拟标签的函数
-	//std::cout << "TagFun()..." << endl;
 	Response_Info res = TagFun(ri);
 
 	// 调用阅读器的函数
-	//std::cout << endl << "ReaderFun()..." << endl;
 	Result_Info result = ReaderFun(res, R1);
 
-	timer.Stop();			// 计时器终止
-
-	//cout << endl << "Result: " << endl;
-	//if (result == TRUE)
-	//	cout << "\tSuccess! Tag is authentication." << endl;
-	//else cout << "\tFailure" << endl;
-	//cout << endl << "Test end." << endl;
+	// 计时器终止
+	timer.Stop();			
 
 	return  timer.GetTime();
 }
